backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
pypdf
qdrant-client
sentence-transformers
rank-bm25
langchain
langchain-community
langgraph
langchain-huggingface
langchain-groq
ragas
pandas
tqdm
tiktoken
pytesseract
pillow
unstructured[all-pdf]
clip @ git+https://github.com/openai/CLIP.git